0.5.0 added shouldAllBeInstanceOf, shouldAnyBeInstanceOf, shouldNoneBeInstanceOf matchers
